\name{cloudfil-package}
\alias{cloudfil-package}
\alias{cloudfil}
\docType{package}
\title{
\packageTitle{cloudfil}
}
\description{
\packageDescription{cloudfil}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cloudfil}
\packageIndices{cloudfil}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cloudfil}

Maintainer: \packageMaintainer{cloudfil}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
